sidecar/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
browser-use-sdk>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
sounddevice>=0.5.0
soundfile>=0.13.0
webrtcvad-wheels>=2.0.0
imap-tools>=1.7.0
beautifulsoup4>=4.12.0
